Struct caribon::Parser

pub struct Parser {
    // some fields omitted
}

Parser which can load a string, detects repetition on it and outputs an HTML file

Methods

impl Parser

fn list_languages() -> Vec<&'static str>

Returns a vector containing all languages implemented.

These values are correct values to give to Parser::new.

fn get_ignored_from_string(list: &str) -> Vec<String>

Returns a vector of ignored words from a string.

Arguments

list – A space or comma separated string, containing words that should be ignored (i.e., don't count repetitions on them).

fn get_ignored_from_lang(lang: &str) -> Vec<String>

Returns a vector containing the default ignored words for this language.

fn new(lang: &str) -> Result<Parser>

Returns Ok(Parser) if language is ok, Err(Error) else.

Arguments

lang – The input text language. This will be used to create the stemmer; it also determines what list of ignored words to use.

Example

let result = caribon::Parser::new("english");
assert!(result.is_ok());

let result = caribon::Parser::new("incorrect language");
assert!(result.is_err());

fn with_max_distance(self, max_dist: u32) -> Parser

Sets max distance for repetitions (default 50).

Arguments

max_dist – A number corresponding to a number of words. If two occurences of a same word are separated by more than this distance, it will not be counted as a repetition.

fn with_html(self, html: bool) -> Parser

Sets HTML detection in input (default true).

You should set it to false if a text is text-formatted, and to true if it contains HTML.

fn with_ignore_proper(self, proper: bool) -> Parser

Sets whether repetition detection should ignore proper nouns (default false).

Basically, if set to true, words that start with a capital and are not at the beginning of a sentence won't be counted for repetitions. Currently, there are still counted if they are in the beginning of a sentence, but with most texts it won't be enough to highligth them as repetitions.

fn with_ignored(self, list: &str) -> Parser

Sets the ignored list with a list of words contained in the argument string.

Arguments

list – A comma or whitespace separated list of words that should be ignored.

fn tokenize(&self, s: &str) -> Result<Vec<Word>>

Tokenize a string into a list of words.

This is the step that converts a string to some inner representation.

Arguments

s – The string to tokenize.

fn detect_leak(&self, vec: Vec<Word>, threshold: f32) -> Vec<Word>

Detect the local repetitions using a leak value.

Basically, each time a word occurs, increase value by 1.0 and each time it does not, multiply by leak (default: 0.98).

This algorithm give results that are not as good than detect_local: instead of underlining a cluster a repeting words, it will only underline the 'offending' ones, i.e. not the first ones.

On the other hand, this method is faster and should use less memory, so it might be useful for very long texts.

Arguments

vec – A list of words threshold – The threshold to consider a repetition (e.g. 1.5)

fn detect_local(&self, vec: Vec<Word>, threshold: f32) -> Vec<Word>

Detect the local number of repetitions.

For each word, the repetition value is set to the total number of occurences of this word since there has been hat least self.max_distance between two occurences.

It is the default algorithm, and probably the one you want to use.

Arguments

vec – A list of words threshold – The threshold to consider a repetition (e.g. 1.9)

fn words_stats(&self, words: &Vec<Word>) -> (HashMap<String, f32>, u32)

Returns stats about the words

Arguments

words – A reference to a list of words

Returns

This method retuns a tuple: * the first element is a hashmap between stemmed strings and the number of occurences of this word * the second element is the total number of (valid) words in the list (non counting whitespace, HTML tags...)

fn detect_global(&self, vec: Vec<Word>, threshold: f32) -> Vec<Word>

Detect the global number of repetitions.

For each word, repetition value is set to the total number of occurences of this word in whole text, divided by total number of words in the text

Arguments

vec – A vector of Word.

threshold – A threshold to highlight repetitions (e.g. 0.01)

fn words_to_html(&self, words: &Vec<Word>, standalone: bool) -> String

Display the words to HTML, higlighting the repetitions.

Use some basic CSS/Js for underlining repetitions and highlighting the over occurrences of the word under the mouse.

Arguments

words – A vector containing repetitions.

standalone – If true, generate a standalone HTML file.