Struct caribon::Parser

pub struct Parser {
    // some fields omitted
}

Parser type

Methods

impl Parser

fn list_languages() -> Vec<&'static str>

Returns a vector containing all languages implemented

fn get_ignored_from_string(list: &str) -> Vec<String>

Returns a vector of ignored words from a string

Arguments

list – A space or comma separated string

fn get_ignored_from_lang(lang: &str) -> Vec<String>

Returns a vector containing ignored words for this language

fn new(lang: &str) -> Option<Parser>

Returns Some(Parser) if language is ok, None else

fn with_max_distance(self, max_dist: u32) -> Parser

Sets max distance for repetition (default 50)

fn with_html(self, html: bool) -> Parser

Sets HTML detection in input (default true)

fn with_ignore_proper(self, proper: bool) -> Parser

Sets ignore_proper flag (default false)

fn with_leak(self, leak: f32) -> Parser

Sets the leak (default 0.98)

fn with_ignored(self, list: &str) -> Parser

Sets the ignored list to a list contained in the string

Arguments

list – A comma or whitespace separated list of words

fn tokenize(&self, s: &str) -> Vec<Word>

Tokenize a string into a list of words

Arguments

s – The string to tokenize

fn detect_leak(&self, vec: Vec<Word>) -> Vec<Word>

Detect the local repetitions, using a leak value

Basically, each time a word occurs, increase value by 1.0 and each time it does not, multiply by leak (default: 0.98)

fn detect_local(&self, vec: Vec<Word>) -> Vec<Word>

Detect the local number of repetitions.

For each word, value is set to the total number of occurences of this word but it is reset to zero if there is more than self.max_distance between two occurences.

fn detect_global(&self, vec: Vec<Word>, is_relative: bool) -> Vec<Word>

Detect the global number of repetitions.

For each word, value is set to the total number of occurences of this word.

Arguments

vec – A vector of Word

is_relative – If true, divide the number of occurences by total number of words

fn words_to_html(&self, words: &Vec<Word>, threshold: f32, standalone: bool) -> String

Display the words to HTML

Use some basic CSS/Js for underlining repetitions and highlighting the over occurrences of the word under the mouse.

Arguments

words – A vector containing all words

threshold – The threshold above which words must be highlighted

standalone – If true, generate a standalone HTML file.